requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.14.0
djangorestframework==3.13.1
mysqlclient @ file:///C:/Users/Lenovo/Downloads/mysqlclient-2.1.1-cp310-cp310-win_amd64.whl
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
